pyproject.toml
[project]
name = "mba-api"
version = "1.0.0"
description = "Market Basket Analysis API"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "SQLAlchemy>=2.0.29",
    "mysql-connector-python>=8.3.0",
    "pandas>=2.0.3",
    "numpy>=1.24.4",
    "mlxtend>=0.23.1",
    "scikit-learn>=1.3.2",
    "scipy>=1.10.1",
    "requests>=2.31.0",
    "python-dateutil>=2.9.0",
    "prefect>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
